Big files restructuration
 - Changed every pair of (templateName, templatePath) properties to
   the new enum TemplateRef with 2 cases, as we should always have one
   or the other, but never none and never both

 - Moved code from ArgumentsUtils to spead it to the following files:
    - Path+AppSupport.swift
    - Utils.swift
    - OutputDestination.swift
    - TemplateRef.swift
    - Path+Commander.swift
- Grouped all files related to Commander to a dedicated folder
  (Both for better organization of the files by feature as for easily
  identify what files can be included in the Test Target or not)

- Renamed ConfigError and ConfigEntry to subtypes of the Config struct
