+++ Sources/Commander/OutputDestination.swift
@@ -1,0 +1,45 @@
+//
+//  OutputDestination.swift
+//  swiftgen
+//
+//  Created by Olivier HALLIGON on 11/10/2017.
+//  Copyright © 2017 AliSoftware. All rights reserved.
+//
+
+import PathKit
+import Commander
+
+enum OutputDestination: ArgumentConvertible {
+  case console
+  case file(Path)
+
+  init(parser: ArgumentParser) throws {
+    guard let path = parser.shift() else {
+      throw ArgumentError.missingValue(argument: nil)
+    }
+    self = .file(Path(path))
+  }
+  var description: String {
+    switch self {
+    case .console: return "(stdout)"
+    case .file(let path): return path.description
+    }
+  }
+
+  func write(content: String, onlyIfChanged: Bool = false) {
+    switch self {
+    case .console:
+      print(content)
+    case .file(let path):
+      do {
+        if try onlyIfChanged && path.exists && path.read(.utf8) == content {
+          return print("Not writing the file as content is unchanged")
+        }
+        try path.write(content)
+        print("File written: \(path)")
+      } catch let e as NSError {
+        printError(string: "error: \(e)")
+      }
+    }
+  }
+}
